+++ internal/firewall/engine_other.go
@@ -5,6 +5,7 @@ package firewall
 import (
 	"context"
 	"fmt"
+	"net"
 	"time"
 )
 
@@ -93,5 +94,7 @@ func (e *Engine) CleanExpiredSubnets() int                              { return
 func (e *Engine) FlushBlocked() error                                   { return nil }
 func (e *Engine) Status() map[string]interface{}                        { return nil }
 
-func (e *Engine) UpdateCloudflareSet(_, _ []string) error { return nil }
-func (e *Engine) CloudflareIPs() ([]string, []string)     { return nil, nil }
+func (e *Engine) UpdateCloudflareSet(_, _ []string) error        { return nil }
+func (e *Engine) CloudflareIPs() ([]string, []string)            { return nil, nil }
+func (e *Engine) SetDOSExemptProviderNets(_ []*net.IPNet)        {}
+func (e *Engine) RefreshDOSExemptSets(_ []*net.IPNet) error      { return nil }